Move Sidebar into separate directory
